Add media info modal template
The media info modal template will be used for the modal when the user
clicks on "Info" button in the standard button group for either Movies
or TV Shows (and possibly Music). In order to use the info modal,
inheriting templates should:

- Extend the 'info.html' template

- Override the 'content' block to provide info in the appropriate
  fashion for the media type.

The template is intended to be used inside of an iframe inside the
bootstrap modal. While it's width should be fixed to the configured
width of the parent document, the height can vary based on the
inheriting template's needs.