Add and enable movie info modal
The movie info modal is enabled for both the index and movies page.

The movie-info template extends the 'info.html' template and provides
information for the user on the movie for which the 'info' button in the
standard button group is clicked. This modal is loaded inside of an
iframe inside of the bootstrap modal 'infoModal'. The iframe will
automatically adjust in size based on the internal contents. As with the
parent template, the width must be fixed but the height can vary.

In this case, the height will vary, mostly based on the length of the
plot information for the movie.

The corresponding python function, info_movie, grabs info using the
GetMovieDetails JSON-RPC call and processes the info for display in a more
human readable format, a process that in this case is more efficient when
done in Python rather than be limited by the templating language.