Create pages for browsing by title and genre
A few templates added to enable browsing by title and genre -- there are
two styles for browsing by title: List mode, and media info mode.

The two methods for listing movies allow for the basics in browsing by
title. The template for genres is used both for selection of a genre and for
selection of a movie within a genre.

The bottom of the page can be covered by the bottom navbar -- this was
resolved by setting a padding on the bottom of the "row" bootstrap
class, putting more space on the bottom of the rendered page.