Add front page for TV show browsing
The front page will allow for browsing through TV shows by title and
also, like the movies page, shows a grouping of the most recent episodes
added to Kodi.

The recently added section follows the same template as the recently
added movies section as found on the main movies page.

This makes progress on #7 but doesn't allow the ability for browsing
through seasons and episodes, which remains to be completed.