Enable viewing seasons of TV Shows
Following the same patterns as many of the other info pages, the seasons
can be viewed in the left panel with a selection of recent episodes in
the media info panel on the right.

The tv_show_seasons() call makes three separate JSON-RPC calls to return
information to the template:

- GetTVShowDetails call for details of the show, likely to be more fully
  utilized in the future.

- GetEpisodes call for the 15 most recent episodes of the TV show. This
  requires a specification of ordering.

- GetSeasons call to actually find the available seasons of the TV show.
  Only two pieces of information are necessary from this call, the label
  of the season and its season id.

This makes progress on #7, leaving only a method for navigating episodes
within a season to finish out the TV show browsing requirement.