Show episodes within a season
Users will be able to select a season of a television show and see
the episodes within that season.

The episode display is the standard media info display. In further
development, this should be able to be selected in a list view much like
the full movies view.

The python function, tv_show_seasons_episodes performs two RPC calls,
one to receive the details of the TV show and another to retreive the
episode information for the season in question.