Add volume slider to footer
The volume slider is very similar to the progress bar slider, and is
updated by the same timer and method updateSlider() as the progress bar.

The volume is provided by a different RPC call than the call used for
the progress information, so a second call is made,
Application.GetVolume(), which gets the volume as an integer and the
mute status as a boolean. The two are combined and returned as JSON for
client-side processing.

To prepare for future additions, some spacing and filler information has
been added to the footer which isn't directly related to the volume
control.