Add play and pause buttons to footer
We add another property to the get_properties RPC call which gets the
"speed" property from Kodi. A speed of 0 represents a paused state, so
we use this value to manage the play vs pause icon.

The remote has calls for stop and play/pause and we use those HTTP calls
again to simplify the implementation. The remainder of the
implementation difficulty is managing the proper icon to display.

Closes #13