Display text when image is unavailable in TV pages
This makes unavailable images show a textual representation of the show
or episode rather than a blank image. This has previously been in place
on the index page and is now extended into the television section.

The image is generated on the fly using the holder.js script which
generates a placeholder image. The placeholder is displayed behind all
images and will show through when there is no image to display.

Closes #11