Get currently playing item in get_properties
We add another RPC call, GetItem(), which returns the current playing
item in Kodi. This will be returned in the get_properties call, and
returned in JSON to be processed on the client side with JavaScript.

The player option sent in the RPC call is the video player and will be
adjusted later on when audio playing is implemented.