Display TV show title near progress bar
We take the currently playing item from the get_properties call as added
in dd5f89b and do further processing in JavaScript to display the title.

The items are truncated at a length of 22 characters, otherwise the
footer bar will need to grow to accommodate the size. The user can
rollover and read the full title if necessary.

Makes progress on #14, however it will only work for TV Shows and should
work across the board for all types of media.