Set action for search form control
We add the action attribute to the search form on the top navbar. This
points to the /search URL, where the search results will be returned.
This commit is in pursuit of goals described in #17 and #20.

We also adjust the initial document focus to attach to the id
"search-input" rather than the class "form-control" because the class
could potentially occur more than once at some point.