Enable search results for movies
A method, search_results() facilitates the process of getting search
results. After getting a list of movies (by title and ID) which match
the result, the method makes the appropriate RPC calls to get details
for each of the search results. This is much faster than the alternative
of getting all details for all movies through RPC, and then picking
the matching items.

The search-results template follows the formats in the browsing sections
and displays a movie list of the matching movies.

This is at most one-half of the results that we plan to return on this
template. In the future, TV shows (#20) and other media will show among
these results

Closes #17