Enable search results for TV shows
Again, as in 3de0daa, search_results() facilitates the process of
getting search results. After getting a list of TV shows
(by title and ID) which match the result, the method makes the
appropriate RPC calls to get a list of seasons for the TV show.

The search results for TV shows includes a link to the TV show page as
well as a link to the available seasons in the database.

Closes #20