Limit search terms to two or more characters
Two characters is the minimum reasonable search term, however most
likely not useful. In this case we allow for the user to make sense of
what they want. We disable one character searches because they basically
return every single result in the database and can lead to minute-long
queries in larger databases.