Allow now playing info to display movie titles
We adjust the JavaScript for the title of the currently displaying item
to support movie titles. This occurs in the same fashion as TV shows as
introduced in 2d9c03a.

This covers the main media types: movies and TV shows. When music and
other are covered, they will be under a separate issue.

Closes #14