Show titles for unknown media types
Streaming YouTube videos or files within the library which
do not have embedded information often have the media type
(as given in get_properties) of "unknown".

We now differentiate between a media type of "unknown" and a
null media type by choosing to display the title information
for "unknown" as we would other types of media. This is
actually rather safe since the JS will interpret a non-existent
title as an empty string, therefore showing no title at all.

Closes #23