Add check marks for played episodes
For places where episodes are listed, we add a check mark to show
whether or not the episode has been watched. This comes directly from
the Kodi database and TARS displays the mark as either on or off - no
other processing is done.

In addition, a title attribute is given to the mark that shows the play
count for the item (as given by the database). This enables the user to
roll-over the check mark to quickly see the number of times the item has
been played.