Separate KODI_URI setting and JSONRPC_URI setting
We have previously used KODI_URI in both contexts, however this may not
be the most ideal practice, as this makes the assumption that TARS can
access the JSON-RPC API and the web browser client can access the Kodi
image web server at the same URI. There are some system set ups in which
this may not be the case.

From this point, two settings are required, although for most simple
Kodi installations, the two will likely be set the same address. They
are:

- JSONRPC_URI: The URI, without the 'jsonrpc' string, where
  TARS can make requests to Kodi's JSON-RPC API

- KODI_URI: The URI where TARS web browsing clients can make requests to
  retreive images from Kodi's web server.