Enable resuming of movies to resume time
As part of the broader goal of #24, we make significant progress in
enabling the user to resume a movie to the resume time stored in the
Kodi database.

This is accomplished through a request to the server when 'play' is
clicked checking if a resume time exists. If not, the item plays
immediately, however, if there is a resume time, the user is presented
with the option of resuming or playing the movie from the beginning.

We introduce two new actions as previously defined in the action-buttons
framework from 166247c. The two action classes are:

- forcePlayMovie
  Used in place of playMovie when no resume check is to be made. Plays
  the movie immediately from the beginning.

- resumeMovie
  A new action which plays the movie but specifically tells Kodi to
  begin at the resume point stored in the database.