Reuse template code across pages when possible
Various portions of the templates are commonly reused, for example, the
action buttons appear consistently and constantly around the templates.
In this case, we have identified the most reused code and simplified it
into templates. In the case of the action buttons we have:

- movie-action-buttons.html
  This is the standard set of horizontal action buttons that is used in
  listings and search results for movies.

- movie-info-action-buttons.html
  This is the standard set of action buttons in two rows that is used in
  the media info view on recent pages and media info listings.

We also have created a template episode-column.html, which provides a
template which is commonly used for TV episodes across various
templates.

This commit should drastically reduce the amount of repeated HTML across
the templates and allow for quicker codebase-wide changes.

Closes #25