Ensure that properties are accessible out of group
The get_properties() method is used for the now playing info bar but
combines a set of several properties for the purposes of efficiency. In
future development, access to some of these properties may be necessary
but the overhead of all of the properties being combined may not be
desired.

The three methods, get_player_properties(), get_app_properties(),
and get_playing_properties() are intended to be used for this situation.
In addition, the methods can be used to hasten server-side development
without needing to specify a set of JSON-RPC requirements.