Test for correct status codes on main pages
We write some tests to ensure that common pages on TARS return proper
status codes. For all of the tests to pass, the JSON-RPC connection must
be available, however most of these tests don't require the connection,
as most of the tests are just looking for a proper HTTP 200 OK response.

To adjust for the tests, we adjust the xbmc object to be instantiated
regardless of how TARS is run, allowing the API to be used in the tests
as well as normal operations.