Test for proper remote functionality in player
We add some tests to ensure that the most-commonly and most important
functions in the remote control trigger the correct action on the Kodi
player.

This requires the tests to have a functioning JSON-RPC connection of its
own, in order to check the state of the player at each point in the
test.

A necessary condition for the tests to proceed is that the Kodi library
contains at least one movie item with the movie_id being 1. This is
likely the case on most Kodi installations, but any installation with an
empty movie library will fail these tests. This movie will, of course,
flash on and off of the Kodi screen while the tests are running.