Test for proper functionality of support functions
We add a class for testing support-type functions in TARS, i.e.,
functions that are present to help format and prepare data to be
consumed in other functions or by the templating engine.

These tests are probably the simplest and most traditional style of unit
testing, where no connection to the JSON-RPC server is required; this is
a just a test that the functions are robust.