Adjust README and docstrings to explain tests
Some instructions are provided in the README to run the tests for TARS.
We also provide some docstrings to the testing classes to explain the
purposes of the testing classes.

The .gitattributes file is changed again to attempt to make github
linguist apply the Jinja2 template highlighting to our template files.