Enable resuming of episodes to resume times
A few matching functions to the movie resuming functions created in
4b61f59 are added to specifically deal with episodes.

Along with the code added for movie resuming in df27f34, two additional
action classes are added to target episode requests:

- forcePlayEpisode
  Used in place of playEpisode when no resume check is made (or after
  the resume check is made to start at the beginning).

- resumeEpisode
  Just like resumeMovie, tells Kodi to start playing the episode at its
  stored time in the Kodi database

Note that the runtime formatting in ask_resume_episode() specifically
casts the response as an integer, owing to the fact that Kodi returns a
floating point number for this but not for any other similar call. We
cast as an integer to ensure that the format_runtime displays integer
numbers. If this is encountered again, we will need to adjust
format_runtime to deal with the problem.

This satisfies all of the requirements in #24, and this commit along
with bf4c970, 4b61f59, and df27f34 allow for resuming across all the
types for which Kodi allows resuming.

Closes #24