Handle TV episodes without images gracefully
Previously when a TV episode didn't have an image available, the info
modal, unlike most areas in TARS, didn't display a placeholder. This
change corrects this.

As part of the correction, we add the commonly used 'holder.js' script
to the 'modal' template, so that it can be used in any modal (which in
this case is primarily the info templates).

Closes #4