Display play icon rather than pause when stopped
This fixes an issue where the pause icon was displayed when the player
is stopped. This was caused by the fact that there is a difference in
the get_properties status between the 'speed=0' approach and when Kodi
reports that nothing is playing at all.

In this case, if Kodi reports that nothing is playing, we force the play
icon to display rather than the pause. This makes sense because if
nothing is playing there is nothing for the user to "pause".

Closes #15