Force resume time to be interpreted as an integer
In the case of shorter movies or tv shows (<1 hour), it appears that
Kodi gives a float value for its resume time rather than the integer
value that is used for longer media. This was previously encountered in
ce780ac when it was thought that it was an oddity without reason, but
would seem there is indeed a reason.

We solve the issue by casting the response from Kodi as an integer. This
doesn't seem to be the case for runtime responses, but this is something
to keep in mind when Kodi responds with a time value in seconds.