Don't display image on resume selection modal
The resume modals had the same issue of handling of an unavailable
image as the information modal did, which was an ungraceful handling
detailed in #4 which was fixed in b62dfc6 and c638070.

Rather than fix the issue in the same way as the information modals, we
remove the image display for the reasons explained below.

The resume modal was made smaller in an effort to differentiate it from
the standard info modal, so the user didn't think they accidentally
clicked for info. Rather than choose to display even smaller images, which
seemed rather comedic when tested, we choose to remove the images
entirely, leaving the resume modal without an image. This further
differentiates the modal and seems to be a better approach.