Show placeholder when movie thumbnail unavailable
On the index and movies pages, movies which did not have a thumbnail
stored would instead display a failed image with the alternate text.
This is an oversight that was missed from earlier development. This
placeholder should show the title of the film in a dynamically generated
holder.js placeholder as done for TV episodes on the index page.

This is related to the issues encountered in #4, however it deals
specifically with listings rather than information modals.