Convert to integer on format_runtime input
To make the format_runtime() function a bit more robust, and since we
encountered problems with the issue in ce780ac and bb9efae, we take all
numeric inputs and convert them to integer. This ensures that later on
all division operations are performed as integer division, which is
obviously important for the function.

This will better handle the random instances in which the JSON-RPC API
returns float values instead of integers for runtimes.