Add .travis.yml continuous integration file
We add the configuration file for Travis CI's continuous integration
system. The necessary requirements are installed and then a selection of
the test suite is run (the tests which do not require a working Kodi
installation). Thus when a build passes things are in pretty good shape
but the full test suite should be run for a greater confidence.