Remove Kodi connection from status code tests
In order to test the status codes of the flask app, the xbmc connection
object is not needed. This is likely a code remnant from when the tests
were brand new and still being imagined.

This change should allow the failed tests experienced in the Travis CI
builds in 213e562 to pass because a Kodi connection is no longer
expected.