Remove status code tests from CI build
Unfortunately the previous approach for testing as described in 1aec40f
was not correct. The status code tests do require a Kodi installation
because they test the TARS searching function, which in turn requires a
JSON-RPC response.

We remove the status code tests and instead we will only test a single
test case for Travis CI builds: TARSSupportFunctionTestCase. This means
that TARS is only lightly tested in the Travis CI builds and further
testing should be done with a working Kodi installation.