Link to show page from now playing indicator
We add to our get_properties and JavaScript main properties loop to
write not only the name of the currently playing TV show but also a link
to the TV show to which the episode belongs. This allows the user to
have quick access to the currently playing show even if navigating
elsewhere.