Adjust testing code to comply with PEP-8 standards
We adjust some of the code in tests.py to better conform to the PEP-8
standards. In this case, we disallow long lines and we clean up some of
the JSON requests to be cleaner and more readable. Pylama confirms that
pep8 runs successfully and the code has no issues.