Show names should be presented as links
If a user, for example, sees the name of a TV show which he/she desires
to watch but not necessarily the most recent episode, it is convienent
to be able to quickly link to that show's page rather than navigate or
serach for it. This remedies that situation

In the process, a CSS class called "noshow" has been created. The
purpose of this class is to allow for links without them obviously
looking like links. In this case, links will be throughout the page and
many underlines and colors would cause distraction.

In addition, we have added to the recent TV episode JSON-RPC queries in
both tv_shows() and index() the addition of 'tvshowid' which is required
to construct the links.