Adjust TARS brand display area to correct link
The brand area display has long displayed 'TARS' but has always been
linked to '#'. While the 'Home' link serves the correct purpose, many
users may be accustomed to clicking a brand image designated in the
upper left corner as a way of 'returning to root'. In this case, we
adjust the link to achieve this functionality.