Add music-related functionality
Back-end items for TARS' music functionality include the modification of
the get_properties function to handle music formats. In addition, a few
remote items that didn't properly handle the audio player were adjusted
to do so. The client JS was updated to properly handle audio information
as well as operate the seek slider.