Add episodes to search
Users may now use the search tool to search through episode names.
Episode names that match the search term will be shown in the search
results.

In addition, now that there are quite a few items in the search results,
we only display their specific result box if there are matching results.